requirements.txt
streamlit==1.32.0
openai==1.12.0
requests==2.31.0
Pillow==10.2.0
chromadb==0.4.22
python-dotenv==1.0.0
transformers==4.38.0
torch==2.2.0
torchvision==0.17.0
streamlit-chat==0.1.1
streamlit-lottie==0.0.5
gTTS==2.5.1
speechrecognition==3.10.3
pydub==0.25.1